pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dcleaderboard"
version = "0.1.0"
description = "Generateur de leaderboard pour le Data Challenge"
readme = "README.md"
packages = [{ include = "dcleaderboard" }]
include = [
	"dcleaderboard/styles.css",
	"dcleaderboard/results/*.json",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pandas = ">=2.0.0"
numpy = ">=1.25,<2.0.0"
matplotlib = ">=3.9.0"
ipython = ">=8.0.0"
pyyaml = ">=6.0.3"
nbformat = ">=5.10.4"
ipykernel = ">=6.30.1"
jupyter = ">=1.0.0"
loguru = ">=0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1"
pytest-cov = "*"
mypy = "*"
ruff = "*"

[tool.poetry.scripts]
dcleaderboard-build = "dcleaderboard.build:main"
